refactor(step1): 1. RacingGame 클래스 리팩토링 - carPositions를 인스턴스 변수 제거 - cars인스턴스 변수 배열 -> arraylist로 변경, 관련된 부분 수정 - move 메소드 리턴값 int[] -> void 수정 / 기존 자동차의 위치정보를 담는 int[]배열값을 리턴하였는데 cars인스턴스변수가 가지고 있기에 리턴할 필요가 없을거라 판단하여 제거 2. ResultView.drawView 메소드 람다식 사용 3. RacingGame 클래스 main메소드 -> RacingGameMain 클래스로 분리
