+++ src/main/java/game/ResultView.java
@@ -1,13 +1,16 @@
 package game;
 
+import java.util.List;
+
 public class ResultView {
-  public static String drawView(int[] positions) {
+  public static String drawView(List<Integer> positions) {
     StringBuilder sb = new StringBuilder();
-    for (int i = 0; i < positions.length; i++) {
-      for (int j = 0; j < positions[i]; j++) {
+    for (int i = 0; i < positions.size(); i++) {
+      int position = positions.get(i);
+      for (int j = 0; j < position; j++) {
         sb.append("-");
       }
-      if (i != positions.length - 1) {
+      if (i != positions.size() - 1) {
         sb.append("\n");
       }
     }
