+++ src/test/java/game/RacingGameTest.java
@@ -1,6 +1,5 @@
 package game;
 
-import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Before;
@@ -13,12 +12,6 @@ public class RacingGameTest {
   @Before
   public void setUp() throws Exception {
     racingGame = new RacingGame(3, 5);
-  }
-
-  @Test
-  public void move() {
-    assertArrayEquals(new int[]{1, 1, 1}, racingGame.move());
-    assertArrayEquals(new int[]{2, 2, 2}, racingGame.move());
   }
 
   @Test
